chore: update typescript packages and script utils
