+++ package.json
@@ -69,7 +69,7 @@
     "typescript": "^5.2.2"
   },
   "resolutions": {
-    "@typescript-eslint/parser": "^6.9.1"
+    "@typescript-eslint/parser": "^6.12.0"
   },
   "packageManager": "pnpm@8.6.10",
   "engines": {
