+++ packages/eslint-config-custom/package.json
@@ -5,9 +5,9 @@
   "private": true,
   "devDependencies": {
     "@dwarvesf/react-eslint-config": "^1.0.1",
-    "@typescript-eslint/eslint-plugin": "^6.9.1",
-    "@typescript-eslint/parser": "^6.9.1",
-    "babel-eslint": "^10.1.0",
+    "@typescript-eslint/eslint-plugin": "^6.12.0",
+    "@typescript-eslint/parser": "^6.12.0",
+    "@babel/eslint-parser": "^7.23.3",
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-typescript": "^17.1.0",
     "eslint-config-next": "^13.2.4",
